% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_band_crosswalk}
\alias{age_band_crosswalk}
\title{Crosswalk to collapse age bands in MSD to match TST for COP23}
\format{
A data frame with 18 rows and 2 variables:
\describe{
  \item{age_msd}{Age bands in the MER structured Dataset}
  \item{age_dp}{collapsed age bands in the TST}
}
}
\usage{
age_band_crosswalk
}
\description{
A dataframe containing the age bands in the MSD and the collapsed age bands in the COP23
Target Setting TOol
}
\keyword{datasets}
