% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_dp.R
\name{agg_dp}
\alias{agg_dp}
\title{Aggregate Targets to IM or PSNU level}
\usage{
agg_dp(df, psnu_lvl = FALSE)
}
\arguments{
\item{df}{data frame to aggregate}

\item{psnu_lvl}{default aggregate is to IM level; if TRUE, aggregates to PSNU level}
}
\description{
Aggregate Targets to IM or PSNU level
}
