% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_msd.R
\name{align_ageband}
\alias{align_ageband}
\title{Align MSD age bands to collapsed age bands in Target Setting Tool}
\usage{
align_ageband(df, collapse = FALSE)
}
\arguments{
\item{df}{dataframe from `align_msd_disagg` to pass through}

\item{collapse}{logical - if TRUE, summarizes across entire dataframe to collapse age bands.
If false, multiple observations with same age bands will occur. Note: if `collapse = TRUE`, processing speed will be long.}
}
\description{
Align MSD age bands to collapsed age bands in Target Setting Tool
}
\examples{
 \dontrun{
   df_final <- align_ageband(df_align, collapse = FALSE)
}

}
