% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_agecoarse.R
\name{align_agecoarse}
\alias{align_agecoarse}
\title{Align Age with Age Coarse}
\usage{
align_agecoarse(df)
}
\arguments{
\item{df}{dataframe from clean_indicators}
}
\description{
Creates a mapping from age to coarse age
}
\keyword{internal}
