% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_class.R
\name{apply_class}
\alias{apply_class}
\title{Apply variable class}
\usage{
apply_class(df)
}
\arguments{
\item{df}{dataframe output to reorder}
}
\description{
Ensure that fiscal year, cumulative, and targets are numeric and all other
variables are stored as characters.
}
