% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_fy}
\alias{apply_fy}
\title{Apply Fiscal Year}
\usage{
apply_fy(df, year)
}
\arguments{
\item{df}{DP dataframe to apply fiscal year to}

\item{year}{fiscal year, derived from `grab_info(filepath, "year")`}
}
\value{
data frame with fiscal year
}
\description{
Apply fiscal year to each row, using the T or T_1 or R in `indicator_code` to
determine whether it's the current or  a prior fiscal year. The fiscal year
can be identified dynamically through `grab_info()`.
}
