% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_prioritization}
\alias{apply_prioritization}
\title{Apply Prioritization}
\usage{
apply_prioritization(df, df_prioritization)
}
\arguments{
\item{df}{Target Setting Tool data frame}

\item{df_prioritization}{dataframe from `grab_prioritization()`}
}
\description{
Join the new COP prioritization onto the target data frame.
}
\seealso{
Other prioritization: 
\code{\link{grab_prioritization}()}
}
\concept{prioritization}
