% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_stamps}
\alias{apply_stamps}
\title{Apply Source File Name and Date Stamp}
\usage{
apply_stamps(df, filepath)
}
\arguments{
\item{df}{data frame read in and reshaped by import_dp and reshape_dp}

\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}
}
\value{
new columns in df with source information
}
\description{
This function applies metadata from the source file to the tidied dataset
including the file name, last modified date, and
}
