% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_inds.R
\name{calculate_inds}
\alias{calculate_inds}
\title{Add Calculated Indicators}
\usage{
calculate_inds(df)
}
\arguments{
\item{df}{dataframe from clean_indicators}
}
\description{
A series of indicators are created from other indicators. This function adds
the following calculated indicators - HTS_TST_POS, OVC_HIVSTAT_D,
PMTCT_STAT_POS, PMTCT_ART.
}
\keyword{internal}
