% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_indicators.R
\name{clean_indicators}
\alias{clean_indicators}
\title{Clean up indicators and disaggs for ease of use}
\usage{
clean_indicators(df)
}
\arguments{
\item{df}{data frame to adjust}
}
\description{
Clean up indicators and disaggs for ease of use
}
