% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_external_mods.R
\name{convert_external_mods}
\alias{convert_external_mods}
\title{Duplicate and convert external modalities to HTS_TST}
\usage{
convert_external_mods(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Duplicate and convert external modalities to HTS_TST
}
