% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Import mechanism specific info from DATIM}
\usage{
get_names(df, map_names = TRUE, psnu_lvl = FALSE, cntry = NULL)
}
\arguments{
\item{df}{data frame to add mechanism info to}

\item{map_names}{import names from DATIM (OU, mechanism, partner) associated with mech_code}

\item{psnu_lvl}{aggregate to the PSNU level instead of IM}

\item{cntry}{country, from grab_cntry() if not connecting to DATIM}
}
\description{
Import mechanism specific info from DATIM
}
