% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_cntry.R
\name{grab_cntry}
\alias{grab_cntry}
\title{Pull OU name from datapack "home" tab}
\usage{
grab_cntry(filepath)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xls}
}
\description{
Function to pull the country name from cell B20 in 'Home' and store as an df
For use later when running `import_plhiv`
}
