% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_info.R
\name{grab_info}
\alias{grab_info}
\title{Pull Information from Data Pack "Home" Tab}
\usage{
grab_info(filepath, type)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xlsx}

\item{type}{either "country" or "year"}
}
\description{
This function extract information stored in the Data Pack Home tab to
identify either the country or what the fiscal year is.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- "../Downloads/DataPack_Jupiter_20200218.xlsx"
cntry <- grab_info(path, "country")
fy <- grab_info(path, "year")
\dontshow{\}) # examplesIf}
}
