% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grab_prioritization}
\alias{grab_prioritization}
\title{Identify Prioritization}
\usage{
grab_prioritization(filepath)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xlsx}
}
\value{
dataframe from the Prioritization tab
}
\description{
Pull from the prioritization tab to have a table of PSNU prioritization for the
current COP.
}
\seealso{
Other prioritization: 
\code{\link{apply_prioritization}()}
}
\concept{prioritization}
