% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grab_snu1}
\alias{grab_snu1}
\title{Identify SNU1 associated with PSNU}
\usage{
grab_snu1(filepath)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}
}
\value{
dataframe from the Prioritization tab
}
\description{
Pull SNU1 from the prioritization tab to have a table to align/apply with the
PSNUxIM tab
}
\seealso{
Other snu1: 
\code{\link{apply_snu1}()}
}
\concept{snu1}
