% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{import_dp}
\alias{import_dp}
\title{Import SNUxIM tab from the Data Pack}
\usage{
import_dp(filepath)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xls}
}
\description{
Initial reading in of the PSNUxIM tab of the Data Pack, which contains all
the data from all the other tabs, broken down by mechanism and PSNU.
}
