% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xlsb.R
\name{import_xlsb}
\alias{import_xlsb}
\title{Import SNU x IM tabs for .xlsb workbooks}
\usage{
import_xlsb(filepath)
}
\arguments{
\item{filepath}{file path to the Data Pack importing}
}
\description{
Import SNU x IM tabs for .xlsb workbooks
}
