% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_file}
\alias{is_file}
\title{Is the filepath correct for the Target Setting Tool}
\usage{
is_file(filepath)
}
\arguments{
\item{filepath}{filepath of Target Setting Tool}
}
\description{
Is the filepath correct for the Target Setting Tool
}
\seealso{
Other validation: 
\code{\link{is_sheet}()},
\code{\link{is_xls}()},
\code{\link{no_connection}()}
}
\concept{validation}
