% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sheet.R
\name{is_sheet}
\alias{is_sheet}
\title{Check if sheet exits in Data pack}
\usage{
is_sheet(filepath)
}
\arguments{
\item{filepath}{filepath of COP 19 Data Pack}
}
\description{
Check if sheet exits in Data pack
}
