% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xls.R
\name{is_xls}
\alias{is_xls}
\title{Check if the filepath is .xls or .xlsx}
\usage{
is_xls(filepath)
}
\arguments{
\item{filepath}{filepath of COP 19 Data Pack}
}
\description{
Check if the filepath is .xls or .xlsx
}
