% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dp_msd.R
\name{join_dp_msd}
\alias{join_dp_msd}
\title{Join TST output with MSD output}
\usage{
join_dp_msd(dp_filepath, msd_filepath)
}
\arguments{
\item{msd_filepath}{}
}
\description{
Join TST output with MSD output
}
