% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lazy_psnuxim_handling}
\alias{lazy_psnuxim_handling}
\title{Lazy PSNUxIM Tool handling}
\usage{
lazy_psnuxim_handling(filepath, type, import_tabs)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}

\item{type}{dataset to extract "PSNUxIM", "SUBNAT", or "ALL" [default] or a specific tab}

\item{import_tabs}{tabs imported as a result of the intersection of
return_tab() and the tabs in the TST}
}
\value{
vector of tabs to read in
}
\description{
The default import option is "ALL" for `tame_dp` which does not include
PSNUxIM due to the inclusion of this tab with the full TST in previous years.
The user will receive an error if they do not switch from the default. This
function switches the type argument in tame_dp from "ALL" to "PSNUxIM" to
avoid this problem and provides the user with both a warning and information
on how to fix the problem running it in the future.
}
\keyword{internal}
