% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_disaggs.R
\name{map_disaggs}
\alias{map_disaggs}
\title{Map Standardized Disaggregate}
\usage{
map_disaggs(df)
}
\arguments{
\item{df}{dataframe from clean_indicators}
}
\description{
To align with DATIM datasets, the standardized disaggregates
for each indicators will be aligned to the Target Setting Tool for FY22 Targets.
}
