% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mer_historic_disagg_mapping_2024}
\alias{mer_historic_disagg_mapping_2024}
\title{Table of MER indicators and disaggs including historic results disaggs from 2024}
\format{
A data frame with 225 rows and 5 variables:
\describe{
  \item{indicator}{MER indicator name}
  \item{numeratordenom}{designates whether the indicator type}
  \item{standardizeddisaggregate}{indicator disaggregation, eg Age/Sex/HIVStatus}
  \item{fiscal_year}{fiscal year}
  \item{kp_disagg}{whether the disaggregation is for Key Populations}
}
}
\source{
\url{https://datim.zendesk.com/hc/en-us/articles/360001143166-DATIM-Data-Entry-Form-Screen-Shot-Repository}
}
\usage{
mer_historic_disagg_mapping_2024
}
\description{
A dataset containing the mapping between MER/SUBNAT/IMPATT indicators from
the Target Setting Tool and their official disaggregates in DATIM from FY24/
COP23 targets, as well as historic results/targets disaggregates from DATIM from
FY22-FY24.
}
\keyword{datasets}
