% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msd_historic_disagg_mapping}
\alias{msd_historic_disagg_mapping}
\title{Table of MER indicators and disaggs including historic results disaggs}
\format{
A data frame with 163 rows and 5 variables:
\describe{
  \item{indicator}{MER indicator name}
  \item{numeratordenom}{designates whether the indicator type}
  \item{standardizeddisaggregate}{indicator disaggregation, eg Age/Sex/HIVStatus}
  \item{fiscal_year}{fiscal year}
  \item{kp_disagg}{whether the disaggregation is for Key Populations}
}
}
\source{
\url{https://datim.zendesk.com/hc/en-us/articles/360001143166-DATIM-Data-Entry-Form-Screen-Shot-Repository}
}
\usage{
msd_historic_disagg_mapping
}
\description{
A dataset containing the mapping between MER/SUBNAT/IMPATT indicators from
the Target Setting Tool and their official disaggregates in DATIM from FY23/
COP22 targets, as well as historic results/targets disaggregates from DATIM from
FY21-FY23.
}
\keyword{datasets}
