% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{no_connection}
\alias{no_connection}
\title{Check if computer has internet connection}
\usage{
no_connection()
}
\description{
Check if computer has internet connection
}
\seealso{
Other validation: 
\code{\link{is_file}()},
\code{\link{is_sheet}()},
\code{\link{is_xls}()}
}
\concept{validation}
