% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ou_ctry_mapping}
\alias{ou_ctry_mapping}
\title{Current Table of PEPFAR Operating Units and Counties}
\format{
A data frame with 51 rows and 2 variables:
\describe{
  \item{operatingunit}{PEPFAR Operating Unit (countries + 3 regional programs)}
  \item{country}{PEPFAR Country Name}
}
}
\source{
\url{https://final.datim.org/api/organisationUnits}
}
\usage{
ou_ctry_mapping
}
\description{
A dataset containing the mapping countries and operating units. Most
countries are also Operating Units, expect for those in regional programs.
}
\keyword{datasets}
