% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_results.R
\name{pivot_results}
\alias{pivot_results}
\title{Pivot Results}
\usage{
pivot_results(df)
}
\arguments{
\item{df}{data frame after it's been aggregated}
}
\value{
data frame with a cumulative column (when/where results exist)
}
\description{
If there are any historic results in the dataset (found in some of the non-
PSNUxIM tabs), we want to separate these from the target values to ensure the
dataset is tidy and results/targets will not be indavertently aggregated. The
reshape will create a cumulative column if results exist in the provided
dataframe.
}
