% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dp_msd.R
\name{prep_dp}
\alias{prep_dp}
\title{Prepare TST tameDP output for join with MSD}
\usage{
prep_dp(dp_filepath)
}
\arguments{
\item{dp_filepath}{TST filepath}
}
\description{
Prepare TST tameDP output for join with MSD
}
