% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_dp.R
\name{reshape_dp}
\alias{reshape_dp}
\title{Reshape Target Setting Tool Long}
\usage{
reshape_dp(df)
}
\arguments{
\item{df}{data frame from import_dp()}
}
\description{
This reshapes the relevant columns from a given tab to long, making it tidy
and more usable. It relies on either `reshape_tab()` or `reshape_psnuim()`
depending on the tab being processed.
}
\seealso{
Other reshape: 
\code{\link{reshape_psnuim}()},
\code{\link{reshape_tab}()}
}
\concept{reshape}
