% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_plhiv.R
\name{reshape_plhiv}
\alias{reshape_plhiv}
\title{Reshape PLHIV Data in Data Pack Long}
\usage{
reshape_plhiv(df)
}
\arguments{
\item{df}{data frame from import_dp()}
}
\description{
This function limits the columns from the Cascade tab to extract PLHIV data
and reshapes it long, so that it is more usable.
}
\seealso{
Other reshape: 
\code{\link{reshape_dp}()}
}
\concept{reshape}
