% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_dp.R
\name{reshape_psnuim}
\alias{reshape_psnuim}
\title{Reshape Target Setting Tool Long}
\usage{
reshape_psnuim(df)
}
\arguments{
\item{df}{data frame from import_dp()}
}
\description{
This function limits the columns from the PSNUxIM tab and reshapes it long,
so that it is more usable. Three values columns are created in the output -
datapacktarget, value, share. This function also splits out the PSNU uid
from the PSNU column.
}
\seealso{
Other reshape: 
\code{\link{reshape_dp}()},
\code{\link{reshape_tab}()}
}
\concept{reshape}
