% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_psnuim.R
\name{reshape_tab}
\alias{reshape_tab}
\title{Reshape Data Pack Tab Long}
\usage{
reshape_tab(df)
}
\arguments{
\item{df}{data frame from import_dp()}
}
\description{
This function limits the columns from a target tab (non PSNUxIM) to extract
data and reshapes it long, so that it is tidy and more usable. This function
also splits out the PSNU uid from the PSNU column.
}
\seealso{
Other reshape: 
\code{\link{reshape_dp}()},
\code{\link{reshape_psnuim}()}
}
\concept{reshape}
