% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_tab}
\alias{return_tab}
\title{Return Tab}
\usage{
return_tab(type)
}
\arguments{
\item{type}{dataset to extract "PSNUxIM", "PLHIV", or "ALL"}
}
\value{
tabs to import
}
\description{
Identify which tab to import based on what you want to use - PSNUxIM, PLHIV,
or ALL.
}
