% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_psnu.R
\name{split_psnu}
\alias{split_psnu}
\title{Clean & Separate PSNU and PSNU UIDS}
\usage{
split_psnu(df)
}
\arguments{
\item{df}{Data Pack data frame from tameDP}
}
\value{

}
\description{
Clean & Separate PSNU and PSNU UIDS
}
