% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{subset_prioritization}
\alias{subset_prioritization}
\title{Subset Prioritization Tab}
\usage{
subset_prioritization(df)
}
\arguments{
\item{df}{data frame after import}
}
\value{
limits to correct columns in data frame from DP tab
}
\description{
Subsets the columns of the massive Data Pack tab down to only those that are
needed. This depends on the type of tab that is being imported. The
Prioritization tab keeps the PSNU and prioritization column.
}
\seealso{
Other subset: 
\code{\link{subset_psnuxim}()},
\code{\link{subset_standard}()}
}
\concept{subset}
