% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{subset_psnuxim}
\alias{subset_psnuxim}
\title{Subset PSNUxIM Tab}
\usage{
subset_psnuxim(df)
}
\arguments{
\item{df}{data frame after import}
}
\value{
limits to correct columns in data frame from DP tab
}
\description{
Subsets the columns of the massive Target Setting Tool tab down to only those that are
needed. This depends on the type of tab that is being imported. PSNUxIM keep
all meta data and taget share/value columns.
}
\seealso{
Other subset: 
\code{\link{subset_prioritization}()},
\code{\link{subset_standard}()}
}
\concept{subset}
