% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{subset_standard}
\alias{subset_standard}
\title{Subset Standard Tabs}
\usage{
subset_standard(df, filepath, tab)
}
\arguments{
\item{df}{data frame after import}
}
\value{
limits to correct columns in data frame from DP tab
}
\description{
Subsets the columns of the massive Data Pack tab down to only those that are
needed. This depends on the type of tab that is being imported. Standard,
non-PSNUxIM/Prioritization) keep column types specified in the Data Pack as
row_header, target, or past.
}
\seealso{
Other subset: 
\code{\link{subset_prioritization}()},
\code{\link{subset_psnuxim}()}
}
\concept{subset}
