% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dp_msd.R
\name{tame_join}
\alias{tame_join}
\title{Join TST output with MSD output}
\usage{
tame_join(tst_filepath, msd_filepath, fy_as_str = TRUE, map_names = FALSE)
}
\arguments{
\item{tst_filepath}{file path to the Target Setting Tool importing, must be .xlsx}

\item{msd_filepath}{filepath to the latest PSNUxIM MSD for corresponding OU}

\item{fy_as_str}{should FY be converted to a string (2025 > FY25) for
Tableau? (default = TRUE)}

\item{map_names}{import names from DATIM (OU, mechanism, partner) associated
with mech_code when working with PSNUxIM (default = FALSE)}
}
\value{
dataframe that combines targets from the TST with corresponding
  historic results/targets from MSD
}
\description{
Join TST output with MSD output
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#TST file path
  tst_path <- "../Downloads/DataPack_Jupiter_20500101.xlsx"
# MSD filepath
  msd_path <- "../Data/MER_Structured_TRAINING_Datasets_PSNU_IM_FY59-61_20240215_v1_1.zip"

#run join function
  df_join <- tame_join(tst_path, msd_path)

#run join function without converting the fiscal year to a string (used in Tableau)
  df_join <- tame_join(tst_path, msd_path, fy_as_str = FALSE)

#run join function with PSNUxIM & map on mechanism info to TST dataframe
  df_join <- tame_join(tst_path, msd_path, map_names = TRUE)
\dontshow{\}) # examplesIf}
}
