% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tame_plhiv.R
\name{tame_plhiv}
\alias{tame_plhiv}
\title{Tidy PLHIV data from Data Pack}
\usage{
tame_plhiv(filepath)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xlsx}
}
\description{
tame_plhiv is a sister function to tame_dp, which readings in the PLHIV data
from the Data Pack and munging in into a tidy data frame to make it more usable to
interact with the data than the way it is stored in the Data Pack. **Given
the changes to the Data Pack each year, the function only works for the
current COP year, COP21.**
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#DP file path
  path <- "../Downloads/DataPack_Jupiter_20200218.xlsx"
#read in data pack
  df_plhiv <- tame_plhiv(path)
\dontshow{\}) # examplesIf}
}
\seealso{
Other primary: 
\code{\link{tame_dp}()}
}
\concept{primary}
