% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tame_subnat.R
\name{tame_subnat}
\alias{tame_subnat}
\title{Export Tidy SUBNAT data from Target Setting Tool}
\usage{
tame_subnat(filepath)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}
}
\description{
tame_subnat is a sister function to tame_dp, which readings in the SUBNAT and
 PLHIV data from the Target Setting Tool and munging in into a tidy data
 frame to make it more usable to interact with the data than the way it is
 stored in the Target Setting Tool. **Given the changes to the Target
 Setting Tool each year, the function only works going back to COP21.**
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#DP file path
  path <- "../Downloads/DataPack_Jupiter_20200218.xlsx"
#read in Target Setting Tool
  df_subnat <- tame_subnat(path)
\dontshow{\}) # examplesIf}
}
\seealso{
Other primary: 
\code{\link{tame_dp}()},
\code{\link{tame_plhiv}()}
}
\concept{primary}
