% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_dp.R
\name{agg_dp}
\alias{agg_dp}
\title{Aggregate Targets to IM or PSNU level}
\usage{
agg_dp(df, psnu_lvl = FALSE)
}
\arguments{
\item{df}{data frame to aggregate}

\item{psnu_lvl}{default aggregate is to IM level; if TRUE, aggregates to PSNU level}
}
\description{
To ensure there are no duplicate rows on the reshape, this function first
aggregates the data by the key columns to minimize any issues. If desiring
to work at the PSNU level, the parameter `psnu_lvl` allows you to aggregate
to the PSNU level instead of the PSNUxIM level.
}
