% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_msd.R
\name{align_msd_disagg}
\alias{align_msd_disagg}
\title{Align MSD extract to disaggregates in Target Setting Tool}
\usage{
align_msd_disagg(msd_path, dp_path, raised_prioritization = FALSE)
}
\arguments{
\item{msd_path}{path to PSNUxIM extract}

\item{dp_path}{path to TST}

\item{raised_prioritization}{logical that is TRUE if OU is setting targets at a raised prioritization level (i.e. SNU1 instead of PSNU)}
}
\description{
This function pulls in a PSNUxIM MSD and datapack filepath to align the MSD extract
to the indicators and disaggregates in the datapack, as well as historic results and targets.
This function also addresses when OUs set targets at a higher level than PSNU for alignment.
}
\examples{
 \dontrun{
   df_msd <- align_msd_disagg(msd_path = msd_path, dp_path = dp_path, raised_prioritization = FALSE)
}

}
