% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_snu1}
\alias{apply_snu1}
\title{Apply SNU1 to dataframe}
\usage{
apply_snu1(df, df_snu1)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}
}
\value{
df Target Setting Tool data frame

df_snu1 dataframe from `grab_snu1()`
}
\description{
Join the SNU1 onto the PSNUxIM data frame.
}
\seealso{
Other snu1: 
\code{\link{grab_snu1}()}
}
\concept{snu1}
