% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dedups.R
\name{convert_dedups}
\alias{convert_dedups}
\title{Adjust Dedup values to negative values}
\usage{
convert_dedups(df)
}
\arguments{
\item{df}{dataframe output to reorder}
}
\description{
The dedup values in the PSNUxIM tab are stored as positive values. This
function converts the dedup mechanism values to negative.
}
