% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Import mechanism specific info from DATIM}
\usage{
get_names(df)
}
\arguments{
\item{df}{data frame to add mechanism info to}
}
\description{
Import mechanism specific info from DATIM
}
