% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Import mechanism specific info from DATIM}
\usage{
get_names(df, ou, map_names, psnu_lvl)
}
\arguments{
\item{df}{data frame to add mechanism info to}

\item{ou}{operating unit, from grab_ou()}

\item{map_names}{import names from DATIM (OU, mechanism, partner) associated with mech_code}

\item{psnu_lvl}{aggregate to the PSNU level instead of IM}
}
\description{
Import mechanism specific info from DATIM
}
