% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{import_dp}
\alias{import_dp}
\title{Import Tabs from the Target Setting Tool}
\usage{
import_dp(filepath, tab)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}

\item{tab}{which sheet to read in}
}
\description{
Initial reading in of tabs of the Target Setting Tool. This function reads in the
necessary tab or tabs, removes unused columns and cleans up the column names
so there are no duplicates. For the PSNUxIM, it identified columns as as a
share or value.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- "../Downloads/DataPack_Jupiter_20200218.xlsx"
df_tst <- import_dp(path, tab = "PSNUxIM")
\dontshow{\}) # examplesIf}
}
