% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plhiv.R
\name{import_plhiv}
\alias{import_plhiv}
\title{Import PLHIV and current on ART from COP20 Datapacks}
\usage{
import_plhiv(filepath)
}
\arguments{
\item{filepath}{cop19 datapack in xlsx format}
}
\description{
Import PLHIV and current on ART from COP20 Datapacks
}
