% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_file.R
\name{is_file}
\alias{is_file}
\title{Is the filepath correct for the Data Pack}
\usage{
is_file(filepath)
}
\arguments{
\item{filepath}{filepath of COP 19 Data Pack}
}
\description{
Is the filepath correct for the Data Pack
}
