% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_sheet}
\alias{is_sheet}
\title{Check if a sheet exits in Data Pack}
\usage{
is_sheet(filepath, tab = "PSNUxIM")
}
\arguments{
\item{filepath}{filepath of Data Pack}

\item{tab}{sheet to check in Data Pack, "PSNUxIM" (default)}
}
\description{
Check if a sheet exits in Data Pack
}
\seealso{
Other validation: 
\code{\link{is_file}()},
\code{\link{is_xls}()},
\code{\link{no_connection}()}
}
\concept{validation}
