% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_xls}
\alias{is_xls}
\title{Check if the filepath is .xls or .xlsx}
\usage{
is_xls(filepath)
}
\arguments{
\item{filepath}{filepath of COP Data Pack}
}
\description{
Check if the filepath is .xls or .xlsx
}
\seealso{
Other validation: 
\code{\link{is_file}()},
\code{\link{is_sheet}()},
\code{\link{no_connection}()}
}
\concept{validation}
