% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{match_col_type}
\alias{match_col_type}
\title{Match Column Type}
\usage{
match_col_type(filepath, tab, pattern = "(row_header|target)")
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xlsx}

\item{tab}{which sheet to read in, "PSNUxIM" (default) or "Cascade" (for PLHIV)}

\item{pattern}{type of column, "assumption", "calculation", "past",
"result", "reference", "row_header", "target"; default = "(row_header|target)"}
}
\value{
Boolean list of matches
}
\description{
Match Column Type
}
