% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dp.R
\name{match_col_type}
\alias{match_col_type}
\title{Match Column Type}
\usage{
match_col_type(filepath, tab, pattern = "(row_header|target|past)")
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}

\item{tab}{which sheet to read in}

\item{pattern}{type of column, "assumption", "calculation", "past",
"result", "reference", "row_header", "target"; default = "(row_header|target|past)"}
}
\value{
Boolean list of matches
}
\description{
This function utilizes the meta data stored in row 6 of each tab of the Data
Pack to determine what column type is - "assumption", "calculation", "past",
 "result", "reference", "row_header", "target". The primary columns we want
 are meta data (row_header), targets, and past (prior year result/targets for
 reference).
}
