% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_dp.R
\name{reshape_dp}
\alias{reshape_dp}
\title{Reshape Data Pack Long}
\usage{
reshape_dp(df)
}
\arguments{
\item{df}{data frame from import_dp()}
}
\description{
This function limits the columns from the PSNUxIM tab and reshapes it long,
so that it is more usable. Three values columns are created in the output -
datapacktarget, value, share.
}
\seealso{
Other reshape: 
\code{\link{reshape_tab}()}
}
\concept{reshape}
