% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_tab}
\alias{return_tab}
\title{Return Tab}
\usage{
return_tab(type)
}
\arguments{
\item{type}{dataset to extract "PSNUxIM", "NAT_SUBNAT" (formerly "PLHIV"),
"ALL", or a specific tab}
}
\value{
tabs to import
}
\description{
Identify which tab to import based on what you want to use - PSNUxIM,
SUBNAT, or ALL (non mechanism tabs). You can also provide a specific tab
name that matches the Target Setting Tool
}
