% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_psnu.R
\name{split_psnu}
\alias{split_psnu}
\title{Clean & Separate PSNU and PSNU UIDS}
\usage{
split_psnu(df)
}
\arguments{
\item{df}{Target Setting Tool data frame from tameDP}
}
\description{
This function removes the contacanated data contained in the same cell. The
psnu column in the Target Setting Tool contains both the psnu, psnuuid, and meta data
on type - Country/SNU/DREAMS/Military. `split_psnu` breaks out psnu and
psnuuid into two columns and removes any other extraneous information.
}
