% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tame_dp.R
\name{tame_dp}
\alias{tame_dp}
\title{Import SNUxIM data from Data Pack and Clean}
\usage{
tame_dp(filepath, map_names = TRUE)
}
\arguments{
\item{filepath}{file path to the Data Pack importing, must be .xls}

\item{map_names}{import names from DATIM (OU, mechanism, partner) associated with mechid}
}
\description{
Import SNUxIM data from Data Pack and Clean
}
\examples{
\dontrun{
#DP file path
  path <- "C:/Users/achafetz/Downloads/DataPack_Malawi_02062019.xls"
#read in data pack
  df_dp <- tame_dp(path) }
}
