% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tame_dp.R
\name{tame_dp}
\alias{tame_dp}
\title{Export Tidy data from Target Setting Tool}
\usage{
tame_dp(filepath, type = "ALL", map_names = FALSE, psnu_lvl = FALSE)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}

\item{type}{dataset to extract "PSNUxIM", "SUBNAT", or "ALL" [default] or a specific tab}

\item{map_names}{import names from DATIM (OU, mechanism, partner) associated with mech_code}

\item{psnu_lvl}{aggregate to the PSNU level instead of IM}
}
\description{
tame_dp is the primary function of the tameDP package, reading in the Data
Pack and munging in into a tidy data frame to make it more usable to
interact with the data than the way it is stored in the Target Setting Tool. **Given
the changes to the Target Setting Tool each year, the function only works for the
current COP year, COP22.**
}
\details{
The main function of `tameDP` is to bring import a COP Target Setting Tool into R
and make it tidy. The function aggregates the COP targets up to the
mechanism level, imports the mechanism information from DATIM, and breaks
out the data elements to make the dataset more usable.

  - Imports Target Setting Tool as tidy data frame
  - Breaks up data elements stored in the indicatorCode column into distinct columns
  - Cleans up the HTS variables, separating modalities out of the indicator name
  - Creates a statushiv column
  - Cleans and separates PSNU and PSNU UID into distinct columns
  - Adds in mechanism information from DATIM, including operatingunit, funding agency, partner and mechanism name
  - Removes any rows with no targets
  - Allows for aggregate to the PSNU level
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#DP file path
  path <- "../Downloads/DataPack_Jupiter_20500101.xlsx"
#read in Target Setting Tool (straight from sheets, not PSNUxIM tab)
  df_dp <- tame_dp(path)
#read in PLHIV/SUBNAT data
  df_dp <- tame_dp(path, type = "SUBNAT")
#read in PSNUxIM data
  df_dp <- tame_dp(path, type = "PSNUxIM")
#apply mechanism names
  df_dp_named <- tame_dp(path, type = "PSNUxIM", map_names = TRUE)
#aggregate to the PSNU level
  df_dp_psnu <- tame_dp(path, type = "PSNUxIM", psnu_lvl = TRUE)
#reading in multiple files and then applying mechanism names (for PSNUxIM)
  df_all <- map_dfr(.x = list.files("../Downloads/DataPacks", full.names = TRUE),
                    .f = ~ tame_dp(.x, map_names = FALSE))
  df_all <- get_names(df_all)
\dontshow{\}) # examplesIf}
}
\seealso{
Other primary: 
\code{\link{tame_plhiv}()},
\code{\link{tame_subnat}()}
}
\concept{primary}
