% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tame_subnat.R
\name{tame_plhiv}
\alias{tame_plhiv}
\title{Export Tidy PLHIV data from Target Setting Tool}
\usage{
tame_plhiv(filepath)
}
\arguments{
\item{filepath}{file path to the Target Setting Tool importing, must be .xlsx}
}
\description{
Deprecated. Use `tame_subnat` instead.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#DP file path
  path <- "../Downloads/DataPack_Jupiter_20200218.xlsx"
#read in Target Setting Tool
  df_subnat <- tame_subnat(path)
\dontshow{\}) # examplesIf}
}
\seealso{
Other primary: 
\code{\link{tame_dp}()},
\code{\link{tame_subnat}()}
}
\concept{primary}
