% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tameDP-package.R
\docType{package}
\name{tameDP-package}
\alias{tameDP}
\alias{tameDP-package}
\title{tameDP: Import PSNUxIM targets from COP19 Data Pack}
\description{
Import PSNUxIM targets from COP20 Data Pack. The purpose is to make it tidy/usable.
}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.gov}

}
\keyword{internal}
