% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tameDP-package.R
\docType{package}
\name{tameDP-package}
\alias{tameDP}
\alias{tameDP-package}
\title{tameDP: Import targets and PLHIV data from COP Target Setting Tool (formerly Data Pack)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Import PSNUxIM targets and PLHIV data from COP Data Pack. The purpose is to make the data tidy and more usable than their current structure in the Excel data packs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usaid-oha-si.github.io/tameDP/}
}

}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.gov}

Authors:
\itemize{
  \item Karishma Srikanth \email{ksrikanth@usaid.gov}
}

Other contributors:
\itemize{
  \item Baboyma Kagniniwa \email{bkagniniwa@usaid.gov} [contributor]
  \item Josh Davis \email{jodavis@usaid.gov} [contributor]
}

}
\keyword{internal}
